collect period shift tests
